Split delete-payload validation into its own module
user-entries.ts already sat near the repo's ~250 LOC guideline before
this PR; adding the delete validator pushed it further over. Move
MAX_DELETE_ENTRY_IDS and validateDeleteUserEntriesPayload (a
self-contained unit with no shared state with the list/update
validators) into user-entries-delete.ts, restoring user-entries.ts to
its pre-PR size.
